# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nutriclock API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/biometric-collection": {
            "get": {
                "tags": [
                    "Biometric Collection"
                ],
                "summary": "Return a list of all biometric collections",
                "description": "Return a list all biometric collections",
                "operationId": "get all biometric collections",
                "responses": {
                    "200": {
                        "description": "return a list of all biometric collections"
                    }
                }
            },
            "post": {
                "tags": [
                    "Biometric Collection"
                ],
                "summary": "Creates new biometric collection",
                "description": "Creates new biometric collection",
                "operationId": "Creates new biometric collection",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "date": {
                                        "type": "string"
                                    },
                                    "intervals": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "groupId": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "return new biometric collection"
                    }
                }
            }
        },
        "/api/biometric-collection/{id}": {
            "delete": {
                "tags": [
                    "Biometric Collection"
                ],
                "summary": "Delete biometric collection",
                "description": "Delete biometric collection",
                "operationId": "Delete biometric collection",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of biometric collection",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return biometric collection"
                    }
                }
            }
        },
        "/api/biometric-group": {
            "get": {
                "tags": [
                    "Biometric Group"
                ],
                "summary": "Return a list of all biometric groups and collections",
                "description": "Return a list all biometric groups and collections",
                "operationId": "get all biometric groups",
                "responses": {
                    "200": {
                        "description": "return a list of all biometric groups and collections"
                    }
                }
            },
            "post": {
                "tags": [
                    "Biometric Collection"
                ],
                "summary": "Creates new biometric group",
                "description": "Creates new biometric group",
                "operationId": "Creates new biometric group",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "return new biometric group"
                    }
                }
            }
        },
        "/api/biometric-group/{id}": {
            "delete": {
                "tags": [
                    "Biometric Group"
                ],
                "summary": "Delete biometric group",
                "description": "Delete biometric group",
                "operationId": "Delete biometric group",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of biometric group",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return biometric group"
                    }
                }
            }
        },
        "/api/biometric-group-user-add": {
            "post": {
                "tags": [
                    "Biometric Collection"
                ],
                "summary": "Associates a biometric group with user",
                "description": "Associates a biometric group with user",
                "operationId": "Associates a biometric group with user",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "userId": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "return biometric group"
                    }
                }
            }
        },
        "/api/biometric-group-user-remove": {
            "post": {
                "tags": [
                    "Biometric Collection"
                ],
                "summary": "Dissociates a biometric group with user",
                "description": "Dissociates a biometric group with user",
                "operationId": "Dissociates a biometric group with user",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "userId": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "return user"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}